Moved Award and Smite component to AwardAndSmiteDialogs folder
