+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardAndSmiteDialogForm/AwardAndSmiteDialogForm.tsx
@@ -33,8 +33,8 @@ import { useTransformer } from '~utils/hooks';
 import { getFormattedTokenValue } from '~utils/tokens';
 import { calculatePercentageReputation } from '~utils/reputation';
 
-import { getUserRolesForDomain } from '../../../transformers';
-import { userHasRole } from '../../../users/checks';
+import { getUserRolesForDomain } from '../../../../transformers';
+import { userHasRole } from '../../../../users/checks';
 
 import { AwardAndSmiteDialogFormValues } from './types';
 import TeamDropdownItem from './TeamDropdownItem';
