+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardAndSmiteDialogForm/TeamDropdownItem.css
