+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardDialog/AwardDialog.tsx
