+++ src/modules/dashboard/components/AwardAndSmiteDialogs/AwardDialog/index.ts
