+++ src/modules/dashboard/components/AwardAndSmiteDialogs/SmiteDialog/SmiteDialog.tsx
