+++ src/modules/dashboard/components/AwardAndSmiteDialogs/SmiteDialog/index.ts
