+++ src/modules/dashboard/components/AwardAndSmiteDialogs/index.ts
@@ -1,0 +1,2 @@
+export { default as SmiteDialog } from './SmiteDialog';
+export { default as AwardDialog } from './AwardDialog';
