+++ src/modules/dashboard/components/ColonyHomeActions/ColonyHomeActions.tsx
@@ -9,8 +9,7 @@ import CreateDomainDialog from '~dashboard/CreateDomainDialog';
 import EditDomainDialog from '~dashboard/EditDomainDialog';
 import CreatePaymentDialog from '~dashboard/CreatePaymentDialog';
 import ManageReputationDialog from '~dashboard/ManageReputationDialog';
-import SmiteDialog from '~dashboard/SmiteDialog';
-import AwardDialog from '~dashboard/AwardDialog';
+import { SmiteDialog, AwardDialog } from '~dashboard/AwardAndSmiteDialogs';
 import ManageDomainsDialog from '~dashboard/ManageDomainsDialog';
 import ManageFundsDialog from '~dashboard/ManageFundsDialog';
 import UnlockTokenDialog from '~dashboard/UnlockTokenDialog';
